Debugged functional derivative of dielectric, now correct
Fixed a bug in the calculation of the convolution, which was not computed
for points with no dielectric, but would not work for the convolution required
in the calculation of the functional derivative: introduced instead a suitable test function.
Fixed a bug in the definition of the functional derivative via finite differences.
Actual code has consistent functional derivatives.