reintroduced the real meaning of boundary
boundary now is the function that goes from 1 to 0 when
passing from the QM to the continuum region.
Switching functions and their derivatives are defined as functions.
Dielectric is built from background and boundary inside dielectric.f90
and not in generate_boundary.f90.
Still need to implement second derivatives and the case of
dielectric regions