substantial modification of main environ driver
Collected the terms depending on boundary, so that kohn-sham potential
and forces can be computed on the same footing. Added a subroutine
to compute partial derivative of boundary with respect to a specific ion.
The dielectric term due to the boundary moved to src/dielectric.f90
going to remove softcavity module. src/cavity.f90 and src/pressure.f90
are almost empty, wondering if they are really needed