substantial modification of input, added BOUNDARY namelist
collected the boundary keywords into a separate namelist. &BOUNDARY
and &ELECTROSTATIC now are only read if there are flags in &ENVIRON
that activate them. Introduced a special env_electrostatic logical keyword
to activate the &ELECTROSTATIC namelist if special electrostatic treatments
are required. environ_type is now processed in a more sensible way,
that works both for SCCS and soft-sphere.