first full implementation of solvent-aware interface
changed keywords names from empty to filled, implemented both the change in
boundary and the change in functional derivative of energy wrt boundary
modified functions so that they can be called with a single function and not just
with arrays. Added a zeroing option to density_of_functions.
Code compiles but still needs to be tested