error due to spurious forces transformed into a warning
the spurious forces due to the core electrons, when using
boundary_mode = 'full', do not cause the calculation to stop, but
only issue a warning. Forces should still be correct, it is just
questionable if the total energy makes sense. It seems that
the error comes up mostly for acidic hydrogens.