substantial restyle of linearized PB, included contribution to forces
charges now include full dielectric and electrolyte, auxiliary charges
of these environments are always computed and updated after the calculation of the potential
to be used in the definition of the energy and of the forces.
Removed auxiliary charges, moved inside dielectric. Gradlog of dielectric is always needed.
Now charges are passed to electrostatic subroutines as INTENT(IN) and are not modified.